requirements.txt
openai
python-dotenv
markdown
pdfkit
